Moves `ImageCache` and `ImageLoaderCacheDecorator` types to production.
